backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
python-dotenv>=1.0.0
# AI & Agents
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-google-genai>=0.0.5
google-genai>=0.3.0
langgraph>=0.0.10
# Data & Graph
neo4j>=5.16.0
chromadb>=0.4.22
python-socketio>=5.11.0
# Perception
easyocr>=1.7.1
pyautogui>=0.9.54
pillow>=10.2.0
